+++ rust/numaflow-core/src/reduce.rs
@@ -10,6 +10,6 @@ mod error;
 /// reading the data it read from ISB or WAL (during replay) to Process and Forward (PNF).
 pub(crate) mod pbq;
 
-mod client;
 /// Process and Forward (PNF) for processing the reduce data and forwarding it to the next vertex.
 mod pnf;
+mod user_defined;
