--- rust/numaflow-core/src/reduce/client.rs
@@ -1,8 +1,0 @@
-pub(crate) mod aligned;
-pub(crate) mod unaligned;
-
-#[derive(Clone)]
-pub(crate) enum ReduceClient {
-    Aligned(aligned::UserDefinedAlignedReduce),
-    Unaligned(unaligned::UserDefinedUnalignedReduce),
-}
