--- rust/numaflow-core/src/reduce/client/unaligned.rs
@@ -1,7 +1,0 @@
-use numaflow_pb::clients::sessionreduce::session_reduce_client::SessionReduceClient;
-use tonic::transport::Channel;
-
-#[derive(Clone)]
-pub(crate) struct UserDefinedUnalignedReduce {
-    client: SessionReduceClient<Channel>,
-}
