+++ rust/numaflow-core/src/reduce/pnf/aligned.rs
@@ -1,0 +1,2 @@
+mod reducer;
+pub(crate) mod user_defined;
