+++ rust/numaflow-core/src/reduce/pnf/aligned/user_defined.rs
@@ -1,7 +1,7 @@
 use crate::Result;
 use crate::config::get_vertex_name;
 use crate::message::{IntOffset, Message, MessageID, Offset};
-use crate::reduce::pnf::{AlignedWindowMessage, WindowOperation};
+use crate::reduce::pnf::aligned::reducer::{AlignedWindowMessage, WindowOperation};
 use crate::shared::grpc::{prost_timestamp_from_utc, utc_from_timestamp};
 use chrono::Utc;
 use numaflow_pb::clients::reduce::reduce_client::ReduceClient;
