+++ rust/numaflow-core/src/reduce/user_defined.rs
@@ -1,0 +1,6 @@
+use crate::reduce::pnf::aligned::user_defined;
+
+#[derive(Clone)]
+pub(crate) enum ReduceClient {
+    Aligned(user_defined::UserDefinedAlignedReduce),
+}
